queue: revise fallback site_name for failed DNS lookups
Previously, we'd use the full scheduled queue name in this case,
which is bad because that can include campaign and tenant information,
which is not really what we want to see in the site name; site names
represent the source -> destination at a more physical level than
the campaign and tenant concept.

Since failed DNS lookups result in messages that cannot be routed,
the site name is less meaningful for this case.

What we do here is take the effective routing domain and prefix
with either NXDOMAIN (for the most likely cause of hitting this code
path) or DNSFAIL (for any other kind of problem with DNS resolution,
which is likely infrastructure related).