ci: remove restore-mtime step
It's purpose was to improve the cache hit rate, which it did,
but it was a dangerous optimization that made the state of the
build sensitive to the date that a given source file was committed,
and made it possible to produce invalid builds if the CI system
is alternating between eg: a PR and the mainline where a given
file was modified at conflicting times.

Dealing with the weirdness is more effort than the optimization
provides as a benefit, so let's just turn this off.