dkim signing: add preliminary support for dkim signing pool
We've been examining a system using http injection where it
appears as though dkim signing might benefit from being broken
out into a dedicated cpu-bound thread pool.

This commit is a fairly quick prototype to allow us to explore
that.

It isn't quite what I'd consider to be the end state for this,
because the pool doesn't export metrics like the others do.
That is because this is a blocking pool rather than an async
pool.

Usage is to define the pool in pre_init:

```
 kumo.on('pre_init', function()
   kumo.dkim.set_signing_threads(24)
 end)
```

If you don't define a pool, then the signing happens
on the calling thread, just as in prior versions.