logging: improve handling for small segment sizes
* When using very small file size or duration constraints, we might
  attempt to create multiple file segments in the same second.  Since
  we require exclusive creation access to the log file name, the
  subsequent attempts to open the segment would fail with a permission
  denied error and cause the associated log record to be dropped.
  Add the number of fractional seconds to the log file name to
  avoid this.

* When using small durations and low traffic, we wouldn't expire log
  files until we had processed 10k records.  That's not so bad in
  a production setting, but in the test harness it is problematic.
  We now check for expiration as part of the file size check so
  that we will prune a segment that is at its time limit.

I don't expect either of these conditions to crop up and matter
in a production setting.

I've added a note to the breaking changes section of the changelog
about the addition of the fractional seconds to the log file name.
I don't expect that to impact anyone in practice either, but I
wanted to call it out as a potential difference in case someone
is using a very precise regex/glob to match the file names.