throttle: increase tolerance of new redis throttle test case
This is acting a bit flakey, and the magnitude expressed as
a percentage is sometimes a bit on the high side at larger
than 10%, so it would be good to understand what is really
going on.

For now we can bump the tolerance so that we're able
to continue producing builds.

refs: https://github.com/KumoCorp/kumomta/issues/297