shutdown: introduce additional stage for shutting down spool
There are certain workloads and traffic patterns that can result
in shutdown taking a long time to complete.  It's not generally
clear to the user what is happening there, so it is desirable
to improve that somehow.

During some recent testing I observed that the rust logic had
completed and that the kumod was process was blocked waiting
for an atexit handler that was joining a rocksdb thread.

This commit introduces an explicit shutdown concept to the spool
abstraction and spool manager.

After we have shutdown all in-flight messages and logs, we now
ask the spool manager to shutdown.  It will steal away the
global refs to the meta and data spools and, concurrently, ask
them to shutdown, and then drop them.

For rocksdb, the shutdown request consists of asking it to
cancel any background work.

For the plain files spool, shutdown is a NOP.

We print out how long it took to perform the shutdown per spool,
as well as indicate when we start to shutdown the spool, as well
as when we are about to return from main. This should help to
understand when a similar atexit shutdown pause is coming into
play in the future.