integration-tests: show logs in sorted order
We were collecting the logs, printing them as we found them, and
then returning the sorted result.

That made it awkward for a human to review, so what we do now
is defer printing the logged records until after we have sorted
them.

For records with the same timestamp, we sort based on the message
id and the record type so that there is some additional consistency
and to make it easier to reason about.