employ a cache around the rustls client config
Since loading the system certificate store can be expensive,
we want to avoid doing it in a hot code path.

This commit introduces a small lru cache based on the relevant
fields of TlsOptions to cache the built client configuration
objects.

In the common case with the default settings, there will just
be a single cached config object that is shared by all sessions.