enhance resolve-site-name script
Adds the ability to pass in --unbound and --dnssec parameters
to influence the resolver setup prior to resolving the site name.

In order to facilitate this, the --script-args handling in kumod
has been given a little facelift: that option has been deprecated
in favor of just collecting an argument list via:

`kumod --script -- arg1 arg2`

which is much more natural and easier to wrangle in the context
of the shell script wrapper that we use for this today.

The old way is preserved for compatibility in case someone is
using this already.

I've also added a requires flag on both the legacy and the
preferred script arguments list to ensure that --script is
used together with it.