Revert "ci: avoid deprecated syntax"
This reverts commit 401df300998096ab208417f4f68b75ef1fdc3d03.

The docker merge step is failing:

```
envconfig.Process: assigning PULLREQUEST_DRONE_PULL_REQUEST to Number: converting '' to type int. details: strconv.ParseInt: parsing "": invalid syntax
```

It seems that the compatibility change in
https://github.com/woodpecker-ci/woodpecker/pull/3939 isn't sufficient.

Let's see if we can still run with the environment set manually,
or if we'll need to blackjack our own woodpecker plugin for this.