try to avoid a startup race with headroom on the CI systems
I noticed that some end to end tests are load shedding, and I think
it is because we initialize headroom to 0 and are racing with
the memory monitor thread startup.

Let's default to non-zero instead.

I think this is manifesting in CI because it has noisy neighbors
on the VM hosts.