treat 3xx responses as transient failures
Following on from b64219bab4c64e9fa20bc68803ec093cf96bfdd5, if we
do encounter a 354 response to a message send, it must indicate
a protocol synchronization issue, so let's treat the associated
message as it had transiently failed and explicitly break the
connection (like we do for a 421).

I did consider making the associated message immediately eligible
to retry, but I worry that if that particular message is a poison
pill we might get stuck in a short cycle of retries for it.

I don't think it is likely for this particular message to be the
poison pill, because it is most likely the victim of an earlier
issue on the connection, but I still want to be cautious here.

The outcome here is still not ideal, but it is better than marking
the associated message as a permanent failure.