tsa: introduce batch-channel in log submission path
This commit introduces a processor pool with n-cpus as
the number of threads as the way that we ingest log
records.

The ingestion is done via a batch-channel which will
allow the potential for automatic batching into larger
transactions in a later commit.

This commit causes publish requests to return as soon
as the record is queued into the channel, rather than
once the record has been processed.  This means that
there is now a window of opportunity where, if tsa-daemon
were to panic or be killed, the associated records will
be forgotten.  There is a maximum of 128k records that
can be in the queue at one time, which puts an upper
bound on this.

In practice, it is not critical if records are lost
in this way: it is not significantly different from
the time window where a node might continue to send
mail before the effect of an automation kicks in.

The net effect of this is that the submission latency
is now much lower while the system remains un-saturated,
but will then block additional submissions until the
backlog is cleared.