tsa: use a single config loading thread
When load testing with 1000 connections and a low nfile ulimit,
I saw that it was possible to get into a state where the majority
of requests would fail due to having too many open files.

This commit addresses that by having a single config loading
thread, which reduces the number of actual files needed to load
the config, which then publishes a shared copy of the resulting
config.  That allows the actual worker threads to grab the most
recent copy and use that, without trying to load it.

If there is a transient issue with reloading the config, the
last published config is retained and used until the next
load attempt succeeds and updates it.